Further refactor around Gateway interface.
