+++ go/vt/vtgate/resolver.go
@@ -41,6 +41,8 @@ var (
 // resharding happened.
 type Resolver struct {
 	scatterConn *ScatterConn
+	toposerv    SrvTopoServer
+	cell        string
 }
 
 // NewResolver creates a new Resolver. All input parameters are passed through
@@ -48,6 +50,8 @@ type Resolver struct {
 func NewResolver(serv SrvTopoServer, statsName, cell string, retryDelay time.Duration, retryCount int, connTimeoutTotal, connTimeoutPerConn, connLife time.Duration) *Resolver {
 	return &Resolver{
 		scatterConn: NewScatterConn(serv, statsName, cell, retryDelay, retryCount, connTimeoutTotal, connTimeoutPerConn, connLife),
+		toposerv:    serv,
+		cell:        cell,
 	}
 }
 
@@ -82,8 +86,8 @@ func (res *Resolver) ExecuteKeyspaceIds(ctx context.Context, sql string, bindVar
 	mapToShards := func(k string) (string, []string, error) {
 		return mapKeyspaceIdsToShards(
 			ctx,
-			res.scatterConn.toposerv,
-			res.scatterConn.cell,
+			res.toposerv,
+			res.cell,
 			k,
 			tabletType,
 			keyspaceIds)
@@ -97,8 +101,8 @@ func (res *Resolver) ExecuteKeyRanges(ctx context.Context, sql string, bindVaria
 	mapToShards := func(k string) (string, []string, error) {
 		return mapKeyRangesToShards(
 			ctx,
-			res.scatterConn.toposerv,
-			res.scatterConn.cell,
+			res.toposerv,
+			res.cell,
 			k,
 			tabletType,
 			keyRanges)
@@ -175,8 +179,8 @@ func (res *Resolver) ExecuteEntityIds(
 ) (*mproto.QueryResult, error) {
 	newKeyspace, shardIDMap, err := mapEntityIdsToShards(
 		ctx,
-		res.scatterConn.toposerv,
-		res.scatterConn.cell,
+		res.toposerv,
+		res.cell,
 		keyspace,
 		entityKeyspaceIDs,
 		tabletType)
@@ -199,8 +203,8 @@ func (res *Resolver) ExecuteEntityIds(
 			resharding := false
 			newKeyspace, newShardIDMap, err := mapEntityIdsToShards(
 				ctx,
-				res.scatterConn.toposerv,
-				res.scatterConn.cell,
+				res.toposerv,
+				res.cell,
 				keyspace,
 				entityKeyspaceIDs,
 				tabletType)
@@ -236,7 +240,7 @@ func (res *Resolver) ExecuteEntityIds(
 // It retries query if new keyspace/shards are re-resolved after a retryable error.
 func (res *Resolver) ExecuteBatchKeyspaceIds(ctx context.Context, queries []proto.BoundKeyspaceIdQuery, tabletType pb.TabletType, asTransaction bool, session *proto.Session) (*tproto.QueryResultList, error) {
 	buildBatchRequest := func() (*scatterBatchRequest, error) {
-		shardQueries, err := boundKeyspaceIDQueriesToBoundShardQueries(ctx, res.scatterConn.toposerv, res.scatterConn.cell, tabletType, queries)
+		shardQueries, err := boundKeyspaceIDQueriesToBoundShardQueries(ctx, res.toposerv, res.cell, tabletType, queries)
 		if err != nil {
 			return nil, err
 		}
@@ -297,8 +301,8 @@ func (res *Resolver) StreamExecuteKeyspaceIds(ctx context.Context, sql string, b
 	mapToShards := func(k string) (string, []string, error) {
 		return mapKeyspaceIdsToShards(
 			ctx,
-			res.scatterConn.toposerv,
-			res.scatterConn.cell,
+			res.toposerv,
+			res.cell,
 			k,
 			tabletType,
 			keyspaceIds)
@@ -316,8 +320,8 @@ func (res *Resolver) StreamExecuteKeyRanges(ctx context.Context, sql string, bin
 	mapToShards := func(k string) (string, []string, error) {
 		return mapKeyRangesToShards(
 			ctx,
-			res.scatterConn.toposerv,
-			res.scatterConn.cell,
+			res.toposerv,
+			res.cell,
 			k,
 			tabletType,
 			keyRanges)
